Decompose {errors} from actions
