+++ src/store/actions/index.js
@@ -1,33 +1,6 @@
 import superagent from "superagent";
+import { errorHandling } from "./errors";
 const baseUrl = "http://localhost:4000";
-
-//add  error:
-export const displayError = payload => {
-  return {
-    type: "ERROR_MESSAGE",
-    payload
-  };
-};
-// remove error:
-export const removeError = () => {
-  return {
-    type: "REMOVE_ERROR"
-  };
-};
-// error handling function
-const errorHandling = async (dispatch, error) => {
-  if (error.response) {
-    await dispatch(removeError());
-    if (error.response.body && error.response.body.message) {
-      const errorMessage = displayError(error.response.body.message);
-      return dispatch(errorMessage);
-    } else if (error.response.body) {
-      const validationError = displayError(error.response.body);
-      return dispatch(validationError);
-    }
-  }
-  console.error(error);
-};
 
 // loading artworks
 export const ARTWORKS_FETCHED = "ARTWORKS_FETCHED";
